pyproject.toml
[tool.poetry]
name = "fastapi-book-catalog"
version = "0.1.0"
description = "FastAPI book catalog"
authors = ["Manuel <manuel.ferreira@ctw.bmwgroup.com>"]
readme = "README.md"
packages = [{ include = "project" }]
 
[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
pyyaml = ">=6.0.2,<7.0.0"
pydantic = ">=2.11.7,<3.0.0"
 
[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"
 
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"